+++ src/django_bird/__init__.py
@@ -2,6 +2,6 @@ from __future__ import annotations
 
 from pluggy import HookimplMarker
 
-__version__ = "0.16.0"
+__version__ = "0.16.1"
 
 hookimpl = HookimplMarker("django_bird")