+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+## [0.16.1]
+
 ### Added
 
 - Added new plugin hook `ready` for handling application initialization during Django's application ready phase.
@@ -405,7 +407,7 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 - Josh Thomas <josh@joshthomas.dev> (maintainer)
 
-[unreleased]: https://github.com/joshuadavidthomas/django-bird/compare/v0.16.0...HEAD
+[unreleased]: https://github.com/joshuadavidthomas/django-bird/compare/v0.16.1...HEAD
 [0.1.0]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.1.0
 [0.1.1]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.1.1
 [0.2.0]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.2.0
@@ -441,3 +443,4 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 [0.14.2]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.14.2
 [0.15.0]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.15.0
 [0.16.0]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.16.0
+[0.16.1]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.16.1